backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.12
pdfplumber==0.11.4
python-docx==1.1.2
groq==0.12.0
langchain==0.3.7
langchain-groq==0.2.1
pydantic==2.9.2
python-dotenv==1.0.1
